Add a function to jump to local symbols
Add a new function, `lsp-ivy-localjump', which jumps to document symbols. As in
VSCode, only the parent symbol is shown with each symbol (unlike `imenu'). Two
sort functions are provided. This function is mainly useful to jump to local
variables.

Optionally, `lsp-ivy-localjump' can display the signatures of symbols
(`lsp-ivy-localjump-signatures'). Because of that, the parent symbol face
(`lsp-ivy-containing-symbol-face') needs to be slightly different from
`lsp-details-face'. Inheriting from `font-lock-type-face' in addition was
chosen.

Since both `lsp-ivy-localjump' and `lsp-ivy--workspace-symbol' need to wrap an
action passing to it the `cdr', drop `lsp-ivy--workspace-symbol-action' and
refactor it into a new function, `lsp-ivy--action', which returns a lambda
wrapping a given CB that way.

Refactor the prefix calculation text calculation into a new function,
`lsp-ivy--prefix', since that is needed both by `lsp-ivy-localjump' and
`lsp-ivy--workspace-symbol'.